refactor(runtime): move request projection to its owner
