+++ packages/runtime/src/ai-sdk-backend.ts
@@ -116,10 +116,11 @@ import {
   type ModelStreamResult,
   type RepairableAiSdkToolCall,
 } from './model-adapter.js';
-import type {
-  RequestProjection,
-  RequestProjectionContext,
-  RequestProjectionStage,
+import {
+  composeRequestProjection,
+  type RequestProjection,
+  type RequestProjectionContext,
+  type RequestProjectionStage,
 } from './request-projection.js';
 import type {
   ActiveToolResultArchiveCandidate,
@@ -234,57 +235,6 @@ export type {
 } from '@maka/core/backend-types';
 
 export const INVALID_TOOL_NAME = 'invalid';
-
-/**
- * Deterministic request-projection pipeline over ONE provider-visible request.
- * Order is a contract: mid-turn capacity compaction runs first among the
- * message-shaping hooks so every later mechanism operates on (and re-converges
- * onto) its projection — active tool-result pruning re-archives large tool
- * results in the rebuilt tail, and semantic/active-full compaction sees the
- * already-compacted messages. On a step where the capacity hook replaced the
- * request, semantic/active-full compaction yields (see send()) so two
- * summarizers never run for one step.
- *
- * Every hook here only SHAPES the projection. The pass/terminate capacity
- * verdict is issued once, after the whole pipeline, by the final-request
- * estimate owner (buildMidTurnFinalRequestVerdict) over the actual outgoing
- * (messages, tools) payload — never by an individual hook over an intermediate
- * projection that a later hook could still rescue.
- */
-export function composeRequestProjection(
-  toolAvailability: RequestProjectionStage | undefined,
-  midTurnCapacityCompact: RequestProjectionStage | undefined,
-  activeToolResultPrune: RequestProjectionStage | undefined,
-  activeFullCompact?: RequestProjectionStage | undefined,
-): RequestProjectionStage | undefined {
-  const hooks = [
-    toolAvailability,
-    midTurnCapacityCompact,
-    activeToolResultPrune,
-    activeFullCompact,
-  ].filter(Boolean) as RequestProjectionStage[];
-  if (hooks.length === 0) return undefined;
-  return async (context: RequestProjectionContext): Promise<RequestProjection | undefined> => {
-    let result: RequestProjection | undefined;
-    let messages = context.messages;
-    for (const hook of hooks) {
-      const hookOptions = {
-        ...context,
-        messages,
-        ...(result?.activeTools ? { activeTools: result.activeTools } : {}),
-      } as RequestProjectionContext;
-      const hookResult = await Promise.resolve(hook(hookOptions));
-      if (!hookResult) continue;
-      result = {
-        ...(result ?? {}),
-        ...hookResult,
-        activeTools: hookResult.activeTools ?? result?.activeTools,
-      };
-      if (hookResult.messages) messages = hookResult.messages;
-    }
-    return result;
-  };
-}
 
 function joinPromptFragments(fragments: readonly (string | undefined)[]): string | undefined {
   const joined = fragments
